Added Trialing_Zeros_in_Factorial.cpp
Returns the number of trailing zeros in a factorial of a given number.
Number of zeros at a end of a given number depends on the number of product of 5*2 are present in it, since quantity of 5 will always exceed the quantity of 2. Find the frequency of 5.

Algorithm:
1. Create a variable count and initialize it with 0.
2. Start a for loop from i = 5, till i < n. where i will be the multiples of 5.
3. For every step. count = count + n/i. This will give the number of 5, 25, 125,....
4. Return count.

>>>
Time complexity = O(n)
Auxiliary space = O(1)